CheckActivity
This is a CheckActivity.class in java folder. This java class is check where your application is run on the first time or not.
If the application is installed and you were open the app first time then this class redirect you Splash Screen and if not then it directly shows MainActivity.